added freebayes module. updated count.py
